transferred db

+++ db/migrate/20190304175055_create_profiles.rb
@@ -2,8 +2,8 @@ class CreateProfiles < ActiveRecord::Migration[5.2]
   def change
     create_table :profiles do |t|
       t.integer :user_id
-      t.integer :starting_weight
-      t.integer :height
+      t.float :starting_weight
+      t.float :height
       t.string :name
       t.string :avatar_image
       t.integer :age
